+++ io_chunks/__init__.py
@@ -1,3 +1,3 @@
-from io_chunks.chunks import RawIOChunk  # noqa: F401
+from io_chunks.raw_io_chunk import RawIOChunk  # noqa: F401
 
 __all__ = ["RawIOChunk"]